Extract workspace index monitoring microcrate
