+++ crates/perl-workspace-index-monitoring/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "perl-workspace-index-monitoring"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Monitoring, limits, and lifecycle instrumentation primitives for workspace index coordination"
+readme = "README.md"
+documentation = "https://docs.rs/perl-workspace-index-monitoring"
+keywords = ["perl", "workspace", "index", "monitoring", "lsp"]
+categories = ["development-tools", "text-editors"]
+include = ["src/**", "tests/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[dependencies]
+parking_lot = "0.12.5"
+
+[lints]
+workspace = true
