+++ crates/perl-workspace-index/src/workspace/monitoring.rs
@@ -1,0 +1,7 @@
+//! Monitoring and lifecycle support types for workspace indexing.
+//!
+//! This module was extracted to the `perl-workspace-index-monitoring` microcrate to
+//! keep workspace indexing focused on symbol extraction/query behavior while
+//! retaining the public path.
+
+pub use perl_workspace_index_monitoring::*;
